// src/app/core/services/model/authorityTemplateNames.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type AuthorityTemplateNames = 'SUPERVISORS' | 'ADMINISTRATORS' | 'USERS' | 'GUESTS' | 'SYSTEM';

export const AuthorityTemplateNames = {
    SUPERVISORS: 'SUPERVISORS' as AuthorityTemplateNames,
    ADMINISTRATORS: 'ADMINISTRATORS' as AuthorityTemplateNames,
    USERS: 'USERS' as AuthorityTemplateNames,
    GUESTS: 'GUESTS' as AuthorityTemplateNames,
    SYSTEM: 'SYSTEM' as AuthorityTemplateNames
};